refactor: React migration: Remove old AnnotationsSlice class
